Refactor code to support JS modular structure and OOP
