+++ src/js/taskList.js
@@ -1,14 +1,11 @@
 import Task from './task.js';
 
 export default class {
-  constructor() {
-    this.tasks = [];
-  }
-
-  getList() {
+  static getList() {
+    const tasks = localStorage.getItem('tasks') ? localStorage.getItem('tasks') : [];
     for (let i = 0; i < 3; i += 1) {
-      this.tasks.push(new Task(i, false, `This Is Task: ${i}`));
+      tasks.push(new Task(i, false, `This Is Task: ${i}`));
     }
-    return this.tasks;
+    return tasks;
   }
 }
